reduce open and close syscalls
Since the fd in our ioctl is only used to get the subvol,
we can keep only one fd per subvol open, and only get the ino for each
file instead of opening them.
This reduce the open (and close) file operation from `ndir + nfile` to
`ndir` because we still open each dir to read_dir.